fixed method call to multiprocessing.cpu_count().
The original code imported the 'processing' module by default and tried to
import the 'multiprocessing' module as a drop in replacement, if processing
could not be found.

In 'processing', there is a method called cpuCount(), while the equivalent
in 'multiprocessing' is called cpu_count().

'multiprocessing' should be used by default here, as it part of the standard
library since 2.6 and has been backported to 2.4 and 2.5.