Make textgrid also run under python3.
This just tweaks the demo print statements so that they will compile
under python3 but still work the same when run under python2.  The
demo output is identical before and after the change, and is identical
between python2 and python3, which means the rest of the code works
fine under python3.